+++ src/test/java/com/mycompany/mytasklist/MyServiceTest.java
@@ -18,7 +18,7 @@ public class MyServiceTest {
         return new LanguageRepository() {
             //przeciazenie metody, aby zawsze zwracala konkretny lang
             @Override
-            Optional<Language> findById(Long id) {
+            Optional<Language> findById(Integer id) {
                 //optional of, a nie ofNullable, bo mamy pewnosc, ze obiekt
                 //istnieje, bo wlasnie go tworzymy
                 return Optional.of(new Language(null, MOCK_WELCOME, null));
@@ -78,7 +78,7 @@ public class MyServiceTest {
         var emptyRepository = new LanguageRepository() {
             //przeciazenie metody, aby nie zwracala jezyka
             @Override
-            Optional<Language> findById(Long id) {
+            Optional<Language> findById(Integer id) {
                 //nie zwracamy zadnego jezyka
                 return Optional.empty();
             }
